Do not store service registry in external resolvers
Currently dependency-management creates a new resolver instance for each repository, for each configuration resolved in a project.
Previously, each repository created a new service registry for providing services to rules like component metadata rules, custom version listers, and custom metadata suppliers.
A service registry is a relatively heavyweight object, and we were not nearly using all the features of a service registry.
Instead, we implement a ServiceLookup, and do so much more efficiently.
We further update DefaultUrlArtifactRepository to avoid instantiating duplicate URI instances for the same url.

Given these changes, in the gradle/gradle build, we go from 106MB to 87MB of DefaultLegacyConfiguration instances, and from 65MB to 54MB of DefaultResolvableConfiguration instances.