Rework TestEventsSpec to be more straightforward
Instead of building up state and verifying ProgressEvents at the end,
we split the verification of the structure and the verification of the
contents into two steps.

This allows us to just worry about finding the appropriate build op and
then simply assert that the content is correct.

We enforce that output and metadata must be checked and we enforce
that all "interesting" progress events are verified. Verified here
means that we look for them and check that they have the correct
parent.