Optimize memory usage of TestTreeModel
This moves all the building of the model into dedicated builder classes,
which then create optimized collections at the end, attempting to
minimize the overhead during copying. This provided a significant
reduction in final memory usage from ~550MB to ~394MB with
`largeMonolithicJavaProject`.

Also optimize TestTreeModel children storage. It turns out we don't really
use the keys here, so making this a list saves a lot of extra references,
and avoids us needing to allocate entries in certain cases.