Address test failures from #35425
The first failure stems from the interaction between gradle.lifecycle.beforeProject and allprojects. We update the build using allprojects to instead write directly to the build files so it can override the beforeProject defaults
The second is due to a removal of a mustRunAfter, affecting parallel tests. The test verified that only a single call to version listing is made, however when executed in parallel multiple calls are made. We add the dependency back to ensure only one task runs at once. Eventually we might want to make the actual implementation resilient to parallel execution, so one request blocks on the other