Rewrite unsafe path name checks, correct dot detection
I needed to exclude '.' from being considered unsafe in the windows checks, but
the amount of repeated searches in the string was getting large. This
replaces the repeated searching with a simple replacement before handing
it off to File/Path APIs for easier-to-understand conditionals.

This does restrict the set of filenames more, as it may now flag
'foo/bar\../baz' as an invalid path on Linux, even though it technically
doesn't contain a traversal.