Use relative path to name FileSource-ed tests in HTML test reports
Changes the test reporting to use the relative file path
instead of the fallback last segment (mangled absolute paths).  This improves clarity.

To enable this, the working dir from the test task is passed to the TestWorker, and ultimately to the test listener via the JUnitPlatformSpec class.

Also cleans up some test expectations now using better names and adds some clarifying comments on the TestDescriptor hierarchy, as those types can't be renamed.