Resolve NPE in resolution engine
Since refactoring strict version constraints in https://github.com/gradle/gradle/issues/35625 some
graph structures can cause NPEs in the resolution engine.

Prior to 35625, there was a bug which erroneously allowed an edge to be added as an outgoing edge
to a node more than once. This was due to improper state tracking in removeOutgoingEdges, causing
us to retain state in `upcomingNoLongerPendingConstraints` even when a node has been removed from
the graph.

This state remained even after being added back to the graph, causing the constraint edge to be
added twice to a given node. In the past, this bug silently caused no problem, however after our
changes to strict versions, we cleared out an edge's selector when the edge was released. Having
two duplicate edges out of a node would cause the duplicate edge to be released twice, causing the NPE.

In practice, an edge should never be added as an outgoing edge to a node more than once. This commit
resolves the duplicate edge bug by ensuring we clean up `upcomingNoLongerPendingConstraints` in
`removeOutgoingEdges`, and document `removeOutgoingEdges` as effectively the inverse of
`visitOutgoingDependenciesAndCollectEdges`.

Since we call cleanupConstraints in removeOutgoingEdges, we refactor existing code which called this
method standalone. Now nodes have two simple operations 'from the outside'. you either want to attach
all outgoing edges, or remove all outgoing edges. It is up to the node to track the internal state,
and no additional public state tracking methods like cleanupConstraints should be necessary.

We additionally strengthen guarantees in the graph, by verifying the reason a component is not selected
when a node is processed. Previously, we had a TODO stating we should determine when a component
is not selected but a node in that component is selected/has incoming edges. This case happens when
a node is part of a module conflict. In these circumstances, we assert that the node is indeed not
connected to the graph, as the conflict handlers are responsible for disconnecting a node's outgoing
edges when it enters a conflict.