Decouple ProjectExecutionServices from mutable Project instance
Currently, nodes in an execution graph can provide their owning project, as context. Services from this
owning project are used to build a ProjectExecutionServices, which then provides services that nodes
can use during execution.

In practice, most services that are required for task execution do not depend on project state. Only a
small subset of services are actually derived from project state.

As part of Isolated Projects, we want to avoid passing around and handling mutable project instances.
By having the task execution services derive from the project's service registry, we run the risk of
access project-scoped services without capturing the correct resources.

This commit updates ProjectExecutionServices to explicitly accept the project-scoped state it needs to be
initialized. This way, we can build the execution services based off of the build scoped services, and allow
the execution services to be customized using the small subset of services that a project uses.

Eventually, we may be able to avoid caching ProjectExecutionServices instances on the project itself, improving
scalability, as we might not necessarily need an instance of these services per project.

In any case, this commit moves us towards the goal of removing references to the Project interface from the execution
engine, so that the engine can live outside of the context of a "project".