Add environment variable support to build options
This support works through the BuildOptionSet which exposes a new
converter for environment variables.
Currently, no subtypes of build option implement `applyFromEnvVar`
except for a no-op implementation in the abstract subtype.

This means that support has to be added per build option.

If usage increases maybe the system should work more like properties.