Cache ancestors strict versions
Make sure to invalidate the cached value if any input to the value changes.

It thea invalidation here is very aggressive. In some contexts like alignment, the invalidation can
clear any caching across the entire graph, effectively nullifying any caching we perform. We may need
to be smarter with invalidation, computing the new value and only invalidating children if the value changes