Configure auto applied Develocity URL through the extension
This allows us to stop using the system property which is a complete
override in the Develocity plugin.

Instead we now only configure the URL passed in if we effectively
auto-applied the plugin. In all other cases, the URL is ignored.

Once the auto applied version of the Develocity plugin understands the
added configuration property, the logic to configure the extension
should be removed - it has a TODO for this.