Attach unattached edges after metadata has been downloaded
In some cases, we might fail to attach an outgoing edge of a node to another node.
When this happens, we remember this edge by tracking it as part of the module that the edge targeted.
Whenever an edge successfully selects a component, we try to find all other unattached edges and attach them.
However, the prior logic did this too early, before the target component's metadata may have been downloaded.
This caused the metadata to be downloaded serially, which may be inefficient.

Instaed, we now attach unattached edges after we successfully attach another edge -- instead of after we successfully select a component version.