Delay attachment of orphaned platform edges
Previously, when we were visiting the dependencies of a virtual platform, we attached any orphaned edges
to the platform immediately. This is the wrong place to attach edges, as we should always attach edges
after selection and metadata downloads.

Instead, we reuse prepareToRecomputeEdge to trigger the orphaned edge to be attached later, at the correct time.