Remove legacy embedded startup logic
Previously, we allowed Gradle to start up from the classpath distribution, which
was a virtual gradle distribution loaded directly from the current classpath.

This is only used during embedded integ tests. However, we also have a 'production'
embedded mode used by TestKit already.

Since we already require that cross version tapi tests already have a full gradle
distribution, we can remove this classpath distribution concept and always funnel
our embedded tapi tests through the production embedded mode.

Also, the in-process gradle executor had some special logic to startup a Gradle invocation in a separate process, skipping
the gradle script. We remove this logic so that the in-process executor uses the normal logic when it is forced
to run gradle invocations in a separate process

Finally, forbid running Gradle without an installation. Now, this is no longer a run scenario we support.

This commit strips out wiring and logic previously used to support running without an installation