Avoid deteched subgraphs in resolved dependency graph
There were a few scenarios where we would short-circuit cleaning up a graph, leading to graphs that contained
detached subgraphs with missing connections between nodes. This graph state surfaced as a corrupt serialized
graph in the legacy ResolvedConfiguration serialized graph. This is because the TransientConfigurationResultsBuilder
serializes all nodes of all components that are part of the graph, even if those nodes are not reachable from
any other node in the graph.

We add additional validation to graphs, ensuring that the source node of all incoming edges to a node is itself
part of the graph, ensuring the final graph of a resolution does not contain disconnected subgraphs.

We also attempted to add additional validation to the graph state, ensuring that all edges are connected to their
source and target nodes correctly, however there are existing bugs in the engine preventing these validations
from passing. We will need to address and un-comment these assertions as part of upcoming work