Do not manually remove outgoing edges of losing component's nodes
The graph automatically does this when it processes a node with no incoming edges. We are doing it too early.
You should not be allowed to change a node's outgoing edges while performing selection, as that can trigger
selection again. When we retarget edges pointing to our losing nodes, we call onFewerSelected, meaning we will
add the losing nodes to the front of the queue. There, their outgoing edges will be removed