This is almost all the way there
One thing we haven't accounted for is the same artifact can be added to different variants.

We have a hack to handle this in one place, but it doesn't work for file collection dependenices added to the root node.
We should instead use a multimap in resolveArtifacts and instead of looping through all sourceVariant/artifact pairs, just loop over all artifacts