Remove final edge case that may cause selection to occur during selection
Whenever an outgoing edge is removed from a node, we release a selector, and may cause module selection to occur.
It is important that nothing that occurs doing module selection can re-trigger module selection, as this can cause reentrant behavior,
which may lead to stack overflows or concurrent modification exceptions.

Since we retarget edges after component replacement during module selection, we need to make sure detaching or attaching an edge to
a target node does not cause us to remove any outgoing edges from a node.

So, we clean up an edge case where we accidentally tried to attach a constraint to a node in a module that was no longer pending.
We fix this by ensuring a module is still no longer pending after we have been notified that a module is not pending and the node
being notified is visited.

This allows us to remove code from attachToTargetNodes that would remove an outgoing edge from a node, and thus removes the possibiltiy
of triggering module selection in this method.