Do not propogate ancestors strict versions from nodes with no incoming edges
When removing an incoming edge from a node, we recompute the ancestors strict versions of that node.
If the node no longer has any incoming edges, it is no longer part of the subgraph and should not
contribute to the ancestors strict versions of its children. However, we do not remove the node's outgoing
edges until it is popped off of the queue, meaning when children eagerly calculate their ancestors strict
versions, the incoming edge from the disconnected parent still contributes to their ancestors strict versions.

So, we ensure we do not use our disconnected parents as part of the contribution of our ancestors strict versions.
Even if we adjusted our approach to be more lazy when we calculated our ancestors strict versions, it is possible
that our parent node is disconnected but will be processed later in the queue than us currently. So, there is no
real way to avoid this, unless we also eagerly disconnected outgoing edges of deselected nodes as opposed to doing
so when it re-enters the queue.

This is something we can investigate in the future -- eagerly deconstructing subgraphs, instead of doing so when the
node comes back around in the queue, though this will take some rearchitecting, as removing outgoing edges from a node
can trigger module version selection.