Permit deferring selection on edges targeting owning virtual platforms
We previously hardcoded false here, and avoided deferring selection on any module that had a virtual platform as the source of a constraint tareting it, as
we previously also had a bug where we treated hard edges targeting a virtual platform as a constraint. Removing that old logic allows us to remove this special case.