Use NDOC<T> as the container type for Kotlin DSL
Instead of using the concrete container subtype, generalize
the accessor to `NDOC<T>`. This is needed because in Kotlin
DSL, the invoke operator for NDOCs has just `NDOC<T>` as the
configuring lambda's receiver and not the more concrete
subtype.

Previously, there was a bug in the DCL schema generation that
led to an NDOC subtype to also introduce the more general
accessor for NDOC<T>. Now it's gone, and this fix for Kotlin
DSL is needed.