Track hidden supertypes (part I)
If a type is a hidden supertype in every type hierarchy that
the DCL schema builder finds, then it should not be imported
into the schema as a data type at all. This prevents the
members from the hidden supertype from still popping up in
schema building (and sometimes causing failures).