Mitigate creating EdgeState more often
Previously, we would only create an EdgeState if we had a hard edge, or a constraint was activated.
Now, after previous changes, we always create an EdgeState even if we have a non activated constraint.
To mitigate this, we push EdgeState creation further down the pipeline, even creating EdgeStates when we have
an excluded dependency. This lets us get rid of two relatively expensive hash maps kept for each node.