Refactor ModuleRegistry
Previously, the ModuleRegistry treated Project modules -- those derived from projects in the gradle build -- from external modules.
External modules would not have a module in the registry, and therefore we were not able to determine their dependencies or other
metadata about the dependency at runtime from the ModuleRegistry.

This commit unifies external dependencies and project dependencies in the ModuleRegistry, treating them as the same concept. Previously,
external dependencies were tracked as part of the 'runtime' classpath of a module and were opaque to the module registy. Now, external
dependencies get a module of their own.

There are two major changes in this commit. First, we change the way we generate the module registry structure. Second, we change the format
of the ModuleRegistry, and thus the structure of the Gradle distribution.

Previously, the metadata of a module was saved in a .properties file within the jar that implemented the module. Since we want to treat external
modules as distribution modules, we want to avoid modifying the jar of external dependencies. So, we now place the properties file corresponding
to a jar next to that jar, where the name of the properties file is the name of the module with the .properties extension. This provides two
efficiency bonuses for the ModuleRegistry. First, the registry does not need to unzip jars in order to determine the structure of the module graph,
and second we can directly lookup modules by name. Previously, we were relying on regexes to map from module name to jar name, as the jar name
includes the version of the module.

The name of the jar is now included in the module properties file. So, previosuly the ModuleRegsitry would find a jar by regex + search, but now it
can directly lookup a module's metadata by name, and the metadata includes the name of the jar.

To generate these properties files for each module, we change from individually generating a metadata file in each jar to generating a set of metadta
files all at once. Previously, the ClasspathManifest task would live in each project, inspect that project's classpath, and write the properties file
to the current project's jar. Now, GenerateClasspathModuleProperties task lives in a distribution project, inspects the entire distribution's classpath
at once, and generates the properties files for all modules in the distribution. This results in significantly fewer tasks, while also allowing the
metadata writing task to consider the entire distribution's classpath at once instead of many sub-classpaths individually. This change is necessary,
as we need to be able to write metadata files for external dependencies, and the previous method did not allow us to do this.

At runtime, when loading modules, we update the ModuleRegistry implementation to be much more efficient. The previous implementation relied on a recusive
graph traveral algorithm which allocated collections at each node in the graph. These allocations were non-negligible for performance, and we had
purpose-built build logic to 'trim' transitive dependencies to make this graph traversal process faster. Now, we implement an iterative graph
traversal algorithm which should be much more efficient. Additionally, it allows traversing starting from multiple root nodes at once, a common
use case in the gradle runtime, improving traversal efficiency, as these roots often contain overlaping transitive subgraphs.

The final notable change is we now store the module coordinates of each distribution module inside its properties file. This allows us to load
the coordinates of the dependency at runtime. So, now the coordinates of our external dependencies are available at runtime, opening the door
to exposing modules from the distribution to users. This was the primary motivation behind this commit. These changes open the door to future changes
that will improve the way dependencies like gradleApi, localGroovy, and toolingApi are loaded from the distribution.

As part of these changes, we extensively reviewed existing logic and determined that the structure of the gradle distribution does not need
to be stable across gradle versions. We did not find evidence of any cross-version bootstrap logic in the tooling API relying on the module
structure of different gradle versions. One exception is DistributionFactory.InstalledDistribution, where the tooling API loads all jars
from the 'lib' directory of a distribution. However, this mechanism does not depend on the module structure -- the properties files -- of
that directory. This commit does not change which jar files are present in the distribution, or where those jar files are located. It just
changes where and what metadata we write about those jar files.